change to cmake parse ascend cann version

+++ lite/backends/nnadapter/nnadapter/driver/huawei_ascend_npu/converter/pad.cc
@@ -25,7 +25,8 @@ int ConvertPad(Converter* converter, hal::Operation* operation) {
 
   // Convert to GE operators
   std::string pad_mode = ConvertPadModeCodeToGEPadMode(mode);
-  auto value = *reinterpret_cast<int32_t*>(value_operand->buffer);
+  auto value =
+      static_cast<int32_t>(*reinterpret_cast<float*>(value_operand->buffer));
   NNADAPTER_CHECK_EQ(pad_mode, "constant")
       << "Only support mode=constant right now, "
          "but received mode is "
